LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-always-https-https-expected.txt

Test that 'unsafe-url' works correctly from HTTPS to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is "https://127.0.0.1:8443/security/contentSecurityPolicy/resources/referrer-test.php?policy=unsafe-url&to=https"
PASS successfullyParsed is true

TEST COMPLETE
